+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Fixed moving and resizing the setting dialog https://github.com/oleg68/GrandOrgue/issues/77
 - `Sample Rate`` and ``Sample Per Buffer`` settings have been moved to the ``Audio Output`` tab
 # 0.3.1.2341-13.os (2021-08-24)
 - Enabled resizing of the main window https://github.com/oleg68/GrandOrgue/issues/71