+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+# 0.3.1.2341-14.os (2021-08-31)
 - Fixed sizing of the setting dialog https://github.com/oleg68/GrandOrgue/issues/77
 - `Sample Rate`` and ``Sample Per Buffer`` settings have been moved to the ``Audio Output`` tab
 # 0.3.1.2341-13.os (2021-08-24)